backend/svc-auth/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
databases[postgresql]==0.7.0
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.2
python-multipart==0.0.6